Add some functions
Add pathjoin() a os.path.join with a normpath
Add incrementString() if there is a number at the end of a string
Add setToList_default() to automatically convert sets to list for json writing. With the help of its "default" argument.
Add openfile() to open a filepath with the default application.
Fix openfolder() to follow python sys.platform doc's guidelines.